measure: Reorganise modules
