+++ eradiate/experiments/_onedim.py
@@ -11,7 +11,7 @@ from ..scenes.atmosphere import Atmosphere, HomogeneousAtmosphere, atmosphere_fa
 from ..scenes.core import KernelDict
 from ..scenes.integrators import Integrator, VolPathIntegrator, integrator_factory
 from ..scenes.measure import DistantMeasure, Measure
-from ..scenes.measure._distant import TargetOriginPoint, TargetOriginSphere
+from ..scenes.measure._target import TargetPoint, TargetSphere
 from ..scenes.surface import LambertianSurface, Surface, surface_factory
 from ..units import unit_context_config as ucc
 
@@ -112,7 +112,7 @@ class OneDimExperiment(EarthObservationExperiment):
                     else:
                         target_point = [0.0, 0.0, 0.0] * ucc.get("length")
 
-                    measure.target = TargetOriginPoint(target_point)
+                    measure.target = TargetPoint(target_point)
 
                 if measure.origin is None:
                     radius = (
@@ -121,7 +121,7 @@ class OneDimExperiment(EarthObservationExperiment):
                         else 1.0 * ucc.get("length")
                     )
 
-                    measure.origin = TargetOriginSphere(
+                    measure.origin = TargetSphere(
                         center=measure.target.xyz, radius=radius
                     )
