+++ eradiate/scenes/measure/tests/test_measure_distant.py
@@ -8,13 +8,12 @@ from eradiate import unit_registry as ureg
 from eradiate._util import onedict_value
 from eradiate.contexts import KernelDictContext
 from eradiate.scenes.core import KernelDict
-from eradiate.scenes.measure._distant import (
-    DistantFluxMeasure,
-    DistantRadianceMeasure,
-    TargetOrigin,
-    TargetOriginPoint,
-    TargetOriginRectangle,
-    TargetOriginSphere,
+from eradiate.scenes.measure._distant import DistantFluxMeasure, DistantRadianceMeasure
+from eradiate.scenes.measure._target import (
+    Target,
+    TargetPoint,
+    TargetRectangle,
+    TargetSphere,
 )
 
 
@@ -23,40 +22,40 @@ def test_target_origin(modes_all):
 
     # TargetOriginPoint: basic constructor
     with ucc.override({"length": "km"}):
-        t = TargetOriginPoint([0, 0, 0])
+        t = TargetPoint([0, 0, 0])
         assert t.xyz.units == ureg.km
 
     with pytest.raises(ValueError):
-        TargetOriginPoint(0)
+        TargetPoint(0)
 
     # TargetOriginPoint: check kernel item
     with ucc.override({"length": "km"}), uck.override({"length": "m"}):
-        t = TargetOriginPoint([1, 2, 0])
+        t = TargetPoint([1, 2, 0])
         assert ek.allclose(t.kernel_item(), [1000, 2000, 0])
 
     # TargetOriginRectangle: basic constructor
     with ucc.override({"length": "km"}):
-        t = TargetOriginRectangle(0, 1, 0, 1)
+        t = TargetRectangle(0, 1, 0, 1)
         assert t.xmin == 0.0 * ureg.km
         assert t.xmax == 1.0 * ureg.km
         assert t.ymin == 0.0 * ureg.km
         assert t.ymax == 1.0 * ureg.km
 
     with ucc.override({"length": "m"}):
-        t = TargetOriginRectangle(0, 1, 0, 1)
+        t = TargetRectangle(0, 1, 0, 1)
         assert t.xmin == 0.0 * ureg.m
         assert t.xmax == 1.0 * ureg.m
         assert t.ymin == 0.0 * ureg.m
         assert t.ymax == 1.0 * ureg.m
 
     with pytest.raises(ValueError):
-        TargetOriginRectangle(0, 1, "a", 1)
+        TargetRectangle(0, 1, "a", 1)
 
     with pytest.raises(ValueError):
-        TargetOriginRectangle(0, 1, 1, -1)
+        TargetRectangle(0, 1, 1, -1)
 
     # TargetOriginRectangle: check kernel item
-    t = TargetOriginRectangle(-1, 1, -1, 1)
+    t = TargetRectangle(-1, 1, -1, 1)
 
     with uck.override({"length": "mm"}):  # Tricky: we can't compare transforms directly
         kernel_item = t.kernel_item()["to_world"]
@@ -72,17 +71,17 @@ def test_target_origin(modes_all):
 
     # TargetOriginSphere: basic constructor
     with ucc.override({"length": "km"}):
-        t = TargetOriginSphere([0, 0, 0], 1)
+        t = TargetSphere([0, 0, 0], 1)
         assert t.center.units == ureg.km
         assert t.radius.units == ureg.km
 
     with pytest.raises(ValueError):
-        TargetOriginSphere(0, 1)
+        TargetSphere(0, 1)
     with pytest.raises(ValueError):
-        TargetOriginSphere([0, 0, 0], -1)
+        TargetSphere([0, 0, 0], -1)
 
     # TargetOriginSphere: check kernel item
-    t = TargetOriginSphere([0, 0, 1], 1)
+    t = TargetSphere([0, 0, 1], 1)
     with uck.override({"length": "mm"}):
         kernel_item = t.kernel_item()
         assert set(kernel_item.keys()) == {"type", "radius", "center"}
@@ -92,25 +91,25 @@ def test_target_origin(modes_all):
 
     # Factory: basic test
     with ucc.override({"length": "m"}):
-        t = TargetOrigin.new("point", xyz=[1, 1, 0])
-        assert isinstance(t, TargetOriginPoint)
+        t = Target.new("point", xyz=[1, 1, 0])
+        assert isinstance(t, TargetPoint)
         assert np.allclose(t.xyz, ureg.Quantity([1, 1, 0], ureg.m))
 
-        t = TargetOrigin.new("rectangle", 0, 1, 0, 1)
-        assert isinstance(t, TargetOriginRectangle)
+        t = Target.new("rectangle", 0, 1, 0, 1)
+        assert isinstance(t, TargetRectangle)
 
     # Converter: basic test
     with ucc.override({"length": "m"}):
-        t = TargetOrigin.convert({"type": "point", "xyz": [1, 1, 0]})
-        assert isinstance(t, TargetOriginPoint)
+        t = Target.convert({"type": "point", "xyz": [1, 1, 0]})
+        assert isinstance(t, TargetPoint)
         assert np.allclose(t.xyz, ureg.Quantity([1, 1, 0], ureg.m))
 
-        t = TargetOrigin.convert([1, 1, 0])
-        assert isinstance(t, TargetOriginPoint)
+        t = Target.convert([1, 1, 0])
+        assert isinstance(t, TargetPoint)
         assert np.allclose(t.xyz, ureg.Quantity([1, 1, 0], ureg.m))
 
         with pytest.raises(ValueError):
-            TargetOrigin.convert({"xyz": [1, 1, 0]})
+            Target.convert({"xyz": [1, 1, 0]})
 
 
 def test_distant_radiance(modes_all):
