+++ eradiate/solvers/onedim/_scene.py
@@ -11,11 +11,8 @@ from ...exceptions import OverriddenValueWarning
 from ...scenes.atmosphere import Atmosphere, HomogeneousAtmosphere, atmosphere_factory
 from ...scenes.core import KernelDict
 from ...scenes.integrators import Integrator, VolPathIntegrator, integrator_factory
-from ...scenes.measure._distant import (
-    DistantMeasure,
-    TargetOriginPoint,
-    TargetOriginSphere,
-)
+from ...scenes.measure._distant import DistantMeasure
+from ...scenes.measure._target import TargetPoint, TargetSphere
 from ...scenes.surface import LambertianSurface, Surface, surface_factory
 from ...units import unit_context_config as ucc
 
@@ -102,7 +99,7 @@ class OneDimScene(Scene):
                     else:
                         target_point = [0.0, 0.0, 0.0] * ucc.get("length")
 
-                    measure.target = TargetOriginPoint(target_point)
+                    measure.target = TargetPoint(target_point)
 
                 if measure.origin is None:
                     radius = (
@@ -111,7 +108,7 @@ class OneDimScene(Scene):
                         else 1.0 * ucc.get("length")
                     )
 
-                    measure.origin = TargetOriginSphere(
+                    measure.origin = TargetSphere(
                         center=measure.target.xyz, radius=radius
                     )
